mat 4 Union.

+++ GameEngineCore/src/maths/maths_func.h
@@ -7,7 +7,7 @@ namespace engine {	namespace maths {
 
 		inline float toRadians(float degrees)
 		{
-			return degrees * (M_PI / 180.0f);
+			return degrees * ((float)M_PI / 180.0f);
 		}
 
 } }
